Module Js.Typed_array

Bindings to the functions in TypedArray.prototype

JavaScript Typed Array API

see https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/TypedArray
MDN

type array_buffer

module ArrayBuffer : sig ... end

The underlying buffer that the typed arrays provide views of

module Int8Array : sig ... end

module Uint8Array : sig ... end

module Uint8ClampedArray : sig ... end

module Int16Array : sig ... end

module Uint16Array : sig ... end

module Int32Array : sig ... end

module Uint32Array : sig ... end

module Float32Array : sig ... end

module Float64Array : sig ... end

module DataView : sig ... end

The DataView view provides a low-level interface for reading and writing multiple number types in an ArrayBuffer irrespective of the platform's endianness.